Cleanup of types
