+++ src/types/chat.ts
@@ -1,20 +1,45 @@
-import { UIUCTool } from '~/utils/functionCalling/handleFunctionCalling'
 import { OpenAIModel } from './openai'
 import { CourseMetadata } from './courseMetadata'
+import { N8NParameter } from './tools'
+
+export interface Conversation {
+  // NO KEY
+  id: string
+  name: string
+  messages: Message[]
+  model: OpenAIModel // ! consider allowing null models: | null
+  prompt: string
+  temperature: number
+  folderId: string | null
+  user_email?: string
+}
 
 export interface Message {
   // id: string;
   role: Role
   content: string | Content[]
   contexts?: ContextWithMetadata[]
-  tools?: ToolResult[]
+  tools?: UIUCTool[]
   responseTimeSec?: number
 }
 
-export interface ToolResult {
-  tool?: UIUCTool
-  toolResult?: string
-  toolContexts?: ContextWithMetadata[]
+export interface UIUCTool {
+  id: string
+  name: string
+  readableName: string
+  description: string
+  parameters?: {
+    type: 'object'
+    properties: Record<string, N8NParameter>
+    required: string[]
+  }
+  courseName?: string
+  enabled?: boolean
+  createdAt?: string
+  updatedAt?: string
+  output?: string
+  loading?: boolean
+  contexts?: ContextWithMetadata[]
 }
 
 // tool_image_url is for images returned by tools
@@ -45,6 +70,7 @@ export interface ContextWithMetadata {
   base_url: string
 }
 
+// These are only internal
 export type Role = 'assistant' | 'user' | 'system'
 
 export interface ChatBody {
@@ -55,18 +81,6 @@ export interface ChatBody {
   isImage: boolean
   courseMetadata?: CourseMetadata
   // NO FOLDER ID
-}
-
-export interface Conversation {
-  // NO KEY
-  id: string
-  name: string
-  messages: Message[]
-  model: OpenAIModel // ! consider allowing null models: | null
-  prompt: string
-  temperature: number
-  folderId: string | null
-  user_email?: string
 }
 
 export interface ChatApiBody {
